requirements.txt
altair==5.0.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
attrs==23.1.0
audioread==3.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
ci-sdr==0.0.2
click==8.1.6
cmake==3.27.0
ConfigArgParse==1.7
ctc-segmentation==1.7.4
Cython==3.0.0
decorator==5.1.1
Distance==0.1.3
editdistance==0.6.2
einops==0.6.1
espnet==202304
espnet-model-zoo==0.1.7
espnet-tts-frontend==0.0.3
fast-bss-eval==0.1.3
filelock==3.12.2
fsspec==2023.6.0
g2p-en==2.1.0
gitdb==4.0.10
GitPython==3.1.32
h5py==3.9.0
huggingface-hub==0.16.4
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==4.13.0
inflect==7.0.0
jaconv==0.3.4
jamo==0.4.1
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
kaldiio==2.18.0
librosa==0.9.2
lit==16.0.6
llvmlite==0.40.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numba==0.57.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
platformdirs==3.10.0
pooch==1.7.0
protobuf==3.20.1
pyarrow==12.0.1
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pydeck==0.8.0
pydub==0.25.1
Pygments==2.15.1
Pympler==1.0.1
pypinyin==0.44.0
python-dateutil==2.8.2
pytorch-wpe==0.0.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyworld==0.3.4
PyYAML==6.0.1
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
resampy==0.4.2
rich==13.5.2
rpds-py==0.9.2
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
sentencepiece==0.1.97
six==1.16.0
smmap==5.0.0
soundfile==0.12.1
streamlit==1.25.0
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.2.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torch-complex==0.4.3
tornado==6.3.2
tqdm==4.65.0
transformers==4.31.0
triton==2.0.0
typeguard==2.13.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
Unidecode==1.3.6
urllib3==2.0.4
validators==0.20.0
watchdog==3.0.0
zipp==3.16.2
